requirements.txt
numpy
pandas
scipy
statsmodels
matplotlib
yfinance
pandas-datareader
python-dateutil
pyyaml
jinja2
tqdm
arch
scikit-learn
